just some mellow weekend bug fixes and updating loot table
-various ingots and gems are now dungeon loot in all valid loot tables, so have fun exploring!
-grinder heads are no longer stackable to avoid issues
-attempted to mabye fix the invisible battery glitch? I have no idea if it worked
-you can now smelt cinnabar in the smelter for mercury
-fixed some recipe errors in the grinder, no mercury stuff yet either
-updated the rock cutter
-fixed an extremely annoying bug where the  show recipe click area overlapped the input slots on the electrolyzer
-dumped some random sprites in a temp folder to use in the future

WARNING - upcoming updates will be messing with world gen, all ores will become stone ore - sand ores will become a seperate variant. then ores will be added to the nether and the end - so some sand ores may or may not be ripped out of the world. Also i will be finally making the workbenches keep their slots so the chance of loosing items in those will be extremely high.