removed ability for full block superconductor cable to be beacon base
this function was originally intended to be paired with the energy siphon because it could draw power from active beacons, but since that is removed there is no need to have full block super conductor cable be beacon compatible. The recipe is also nerfed because they no longer offer any feature other than being a full block, which is a rare use case.